+++ src/commands/mod.rs
@@ -9,4 +9,3 @@ pub mod rb;
 pub mod rm;
 pub mod stat;
 pub mod sync;
-pub mod versions;